refactor: buildMachine goes; the model's Cpu takes a spec and peripherals directly
Once the model names its CPU class, buildMachine was two lines whose only
remaining job was the spec check, and the check was never the point. The
file keeps what is left, machineSpec and nullIo, under a name for that.
