+++ src/fake6502.js
@@ -3,22 +3,21 @@
 import { FakeVideo } from "./video.js";
 import { FakeSoundChip } from "./soundchip.js";
 import { TEST_6502, TEST_65C02, TEST_65C12, tubeModelFor } from "./models.js";
-import { buildMachine, machineSpec, nullIo } from "./build-machine.js";
+import { machineSpec, nullIo } from "./machine-spec.js";
 
 const fakeVideo = new FakeVideo();
 const soundChip = new FakeSoundChip();
 
 export function fake6502(model, opts = {}) {
     model = model || TEST_6502;
-    return buildMachine({
-        model,
-        spec: machineSpec({
+    return new model.Cpu(model, {
+        ...nullIo({ video: opts.video ?? fakeVideo, soundChip: opts.soundChip ?? soundChip }),
+        config: machineSpec({
             tube: opts.tube ? tubeModelFor(model) : null,
             tubeCpuMultiplier: opts.tubeCpuMultiplier,
             cpuMultiplier: opts.cpuMultiplier,
             hasTeletextAdaptor: opts.hasTeletextAdaptor,
         }),
-        io: nullIo({ video: opts.video ?? fakeVideo, soundChip: opts.soundChip ?? soundChip }),
         cycleAccurate: opts.cycleAccurate,
     });
 }
