+++ src/machine-spec.js
@@ -12,8 +12,6 @@ const NullUserPort = {
     },
 };
 
-const SpecBrand = Symbol("machineSpec");
-
 const SpecDefaults = {
     keyLayout: "physical",
     cpuMultiplier: 1,
@@ -32,30 +30,18 @@ const SpecDefaults = {
 /**
  * What a machine is fitted with and how it is driven, complete and frozen:
  * every field is present, an unknown one is an error, and an undefined
- * override means the default. The CPU takes nothing else.
+ * override means the default. It is the `config` a CPU is built with.
  */
 export function machineSpec(overrides = {}) {
     const unknown = Object.keys(overrides).filter((field) => !Object.hasOwn(SpecDefaults, field));
     if (unknown.length) throw new Error(`Unknown machine spec fields: ${unknown.join(", ")}`);
     const given = Object.fromEntries(Object.entries(overrides).filter(([, value]) => value !== undefined));
     return Object.freeze({
-        [SpecBrand]: true,
         ...SpecDefaults,
         ...given,
         extraRoms: Object.freeze([...(given.extraRoms ?? SpecDefaults.extraRoms)]),
         debugFlags: Object.freeze({ ...SpecDefaults.debugFlags, ...given.debugFlags }),
     });
-}
-
-/** Whether `config` was made by machineSpec(), and so is complete and unchanged. */
-export function isMachineSpec(config) {
-    return config?.[SpecBrand] === true;
-}
-
-/** The processor for `model`, fitted as `spec` says, talking to the peripherals in `io`. */
-export function buildMachine({ model, spec, io, cycleAccurate = true }) {
-    if (!isMachineSpec(spec)) throw new Error("A machine's config must come from machineSpec()");
-    return new model.Cpu(model, { ...io, config: spec, cycleAccurate });
 }
 
 /** Peripherals that go nowhere, for a machine run headless. */
